+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-26 16:48:45
-// Commit: perf(test): apply RepaintBoundary on animated test widgets
+// Updated at: 2026-06-26 19:29:50
+// Commit: fix(test): prevent double tap duplicate action in test
 
 class testComponentState {
-  final String commitRef = '2370';
-  final String lastMessage = 'perf(test): apply RepaintBoundary on animated test widgets';
-  final DateTime timestamp = DateTime.parse('2026-06-26T16:48:45+05:00');
+  final String commitRef = '2378';
+  final String lastMessage = 'fix(test): prevent double tap duplicate action in test';
+  final DateTime timestamp = DateTime.parse('2026-06-26T19:29:50+05:00');
 }